data preprocessing scripts added

+++ data_loaders/gesture/scripts/ptbrgesture_prep.py
@@ -30,57 +30,69 @@ def main(args):
         print('Processing wavlm')
         process_wavlm(npy16k, wavlmpath)
 
-def process_wavlm(sourcepath, savepath, split):
+def process_wavlm(sourcepath, savepath):
     wavlm_layer = 11 
     fps=30
     sr=16000
     device = 'cuda' if torch.cuda.is_available() else 'cpu'
     assert os.path.exists(sourcepath), f"audio16k_npy not found in {sourcepath}. Required to process wavlm representations, make sure wav files were processed first."
-    assert os.path.exists(savepath), f"wavlm model directory not found in current directory."
+    #assert not os.path.exists(savepath), f"wavlm model directory already exists."
     if not os.path.exists(savepath):
         os.mkdir(savepath)
     checkpoint = torch.load('./wavlm/WavLM-Base+.pt')
     wavlm_cfg = WavLMConfig(checkpoint['cfg'])
     wavlm = WavLM(wavlm_cfg)
-    #wavlm.to(device)
+    wavlm.to(device)
     wavlm.load_state_dict(checkpoint['model'])
     wavlm.eval()
-    for file in tqdm(os.listdir(sourcepath)):
-        if not os.path.exists(os.path.join(savepath, file)):
-            audio_path = os.path.join(sourcepath, file)
-            # Load with Numpy
-            signal = np.load(audio_path)
-            # Set to model innput format
-            signal = torch.tensor(signal).unsqueeze(0)#.to(device)
-            # Normalize
-            if wavlm_cfg.normalize:
-                signal_norm = torch.nn.functional.layer_norm(signal , signal.shape)
-            else:
-                signal_norm = signal
-            # Run Model (rep=Desired Layer, layer_results=all layers)
-            rep, layer_results = wavlm.extract_features(signal_norm, output_layer=wavlm_layer, ret_layer_results=True)[0]
-            layer_reps = [x.transpose(0, 1) for x, _ in layer_results] # fix shape
-            # Get Number of Seconds of Audio File
-            n_secs = signal.shape[1] / sr
-            # Get Number of poses equivalent to audio file duration, given fps (alignment len)
-            n_pose = n_secs * fps
-            # Interpolate number of representations to match number of poses corresponding to audio file
-            interp_reps = F.interpolate(rep.transpose(1, 2), size=int(n_pose), align_corners=True, mode='linear')
-            # Prepare to save
-            interp_reps = interp_reps.squeeze(0).transpose(0,1).cpu().detach().data.cpu().numpy()
-            # Double check dimension
-            assert (interp_reps.shape[0] == int(np.ceil(n_pose)) or interp_reps.shape[0] == int(np.floor(n_pose)))
-            np.save(os.path.join(savepath, file), interp_reps)
+    with torch.no_grad():
+        for file in tqdm(os.listdir(sourcepath)):
+            if not os.path.exists(os.path.join(savepath, file)):
+                audio_path = os.path.join(sourcepath, file)
+                # Load with Numpy
+                signal = np.load(audio_path)
+                if signal.shape[0] < 960000: #1 minute
+                    interp_reps = getwavlmrep(signal, wavlm, device, wavlm_layer, wavlm_cfg, sr=sr, fps=fps)
+                else: #Break the file into smaller chunks to avoid memory issues
+                    interp_reps = []
+                    for subsignal in np.array_split(signal, np.ceil(signal.shape[0]/960000)):
+                        subinterp_reps = getwavlmrep(subsignal, wavlm, device, wavlm_layer, wavlm_cfg, sr=sr, fps=fps)
+                        interp_reps.append(subinterp_reps)
+                    interp_reps = np.vstack(interp_reps)
+                np.save(os.path.join(savepath, file), interp_reps)
+
+def getwavlmrep(signal, wavlm, device, wavlm_layer, wavlm_cfg, sr=16000, fps=30):
+    # Set to model innput format
+    signal = torch.tensor(signal).unsqueeze(0).to(device)
+    # Normalize
+    if wavlm_cfg.normalize:
+        signal_norm = torch.nn.functional.layer_norm(signal , signal.shape)
+    else:
+        signal_norm = signal
+    # Run Model (rep=Desired Layer, layer_results=all layers)
+    rep, layer_results = wavlm.extract_features(signal_norm, output_layer=wavlm_layer, ret_layer_results=True)[0]
+    layer_reps = [x.transpose(0, 1) for x, _ in layer_results] # fix shape
+    # Get Number of Seconds of Audio File
+    n_secs = signal.shape[1] / sr
+    # Get Number of poses equivalent to audio file duration, given fps (alignment len)
+    n_pose = n_secs * fps
+    # Interpolate number of representations to match number of poses corresponding to audio file
+    interp_reps = F.interpolate(rep.transpose(1, 2), size=int(n_pose), align_corners=True, mode='linear')
+    # Prepare to save
+    interp_reps = interp_reps.squeeze(0).transpose(0,1).cpu().detach().data.cpu().numpy()
+    # Double check dimension
+    assert (interp_reps.shape[0] == int(np.ceil(n_pose)) or interp_reps.shape[0] == int(np.floor(n_pose)))
+    return interp_reps
 
 def process_wav(sourcepath, savepath, sr=16000):
-    assert not os.path.exists(savepath), f"audio_16k_npy already exists in {savepath}. Delete it to process again."
+    #assert not os.path.exists(savepath), f"audio_16k_npy already exists in {savepath}. Delete it to process again."
     if not os.path.exists(savepath):
         os.mkdir(savepath)
     for file in tqdm(os.listdir(sourcepath)):
-        #if not os.path.exists(os.path.join(savepath, file[:-4] + '.npy')):
-        signal, _sr = librosa.load(os.path.join(sourcepath, file), mono=True, sr=sr)
-        assert _sr == sr
-        np.save(os.path.join(savepath, file[:-4]+'.npy'), signal)
+        if not os.path.exists(os.path.join(savepath, file[:-4] + '.npy')):
+            signal, _sr = librosa.load(os.path.join(sourcepath, file), mono=True, sr=sr)
+            assert _sr == sr
+            np.save(os.path.join(savepath, file[:-4]+'.npy'), signal)
     return savepath
 
 def process_bvh(bvhpath, rot6dpath, rot3dpath, pospath, takes):
@@ -89,11 +101,12 @@ def process_bvh(bvhpath, rot6dpath, rot3dpath, pospath, takes):
         if not os.path.exists(path):
             os.mkdir(path)
     for file in tqdm(os.listdir(bvhpath)):
-        anim = bvhsdk.ReadFile(os.path.join(bvhpath, file))
-        npyrot6d, npyrot, npypos = bvh2representations1(anim)
-        np.save(os.path.join(rot6dpath, file[:-4]), npyrot6d)
-        np.save(os.path.join(rot3dpath, file[:-4]), npyrot)
-        np.save(os.path.join(pospath, file[:-4]), npypos)
+        if not os.path.exists(os.path.join(rot6dpath, file[:-4] + '.npy')):
+            anim = bvhsdk.ReadFile(os.path.join(bvhpath, file))
+            npyrot6d, npyrot, npypos = bvh2representations1(anim)
+            np.save(os.path.join(rot6dpath, file[:-4]), npyrot6d)
+            np.save(os.path.join(rot3dpath, file[:-4]), npyrot)
+            np.save(os.path.join(pospath, file[:-4]), npypos)
 
 def compute_meanstd(path, savepath, npstep=1, vel=False):
     all_data = []
@@ -124,11 +137,11 @@ def paths_get_and_check(data_dir):
     pospath = os.path.join(motionpath, 'pos')
     npy16k = os.path.join(audiopath, 'npy16k')
     wavlmpath = os.path.join(audiopath, 'wavlm')
-    assert not os.path.exists(rot6dpath), 'rot6d directory already exists'
-    assert not os.path.exists(rot3dpath), 'rot3d directory already exists'
-    assert not os.path.exists(pospath), 'pos directory already exists'
-    assert not os.path.exists(npy16k), 'npy16k directory already exists'
-    assert not os.path.exists(wavlmpath), 'wavlm directory already exists'
+    #assert not os.path.exists(rot6dpath), 'rot6d directory already exists'
+    #assert not os.path.exists(rot3dpath), 'rot3d directory already exists'
+    #assert not os.path.exists(pospath), 'pos directory already exists'
+    #assert not os.path.exists(npy16k), 'npy16k directory already exists'
+    #assert not os.path.exists(wavlmpath), 'wavlm directory already exists'
     return bvhpath, wavpath, rot6dpath, rot3dpath, pospath, npy16k, wavlmpath
     
 def takes_get_and_check(bvhpath, wavpath):
